+++ src/lib/builders/toggle-group/index.ts
@@ -68,13 +68,12 @@ export function createToggleGroup(args: CreateToggleGroupArgs = {}) {
 				disabled?: boolean;
 		  }
 		| string;
-	const item = elementMultiDerived([options, value], ([$options, $value], { createAttach }) => {
+	const item = elementMultiDerived([options, value], ([$options, $value], { attach }) => {
 		return (args: ToggleGroupItemArgs) => {
 			const itemValue = typeof args === 'string' ? args : args.value;
 			const argDisabled = typeof args === 'string' ? false : !!args.disabled;
 			const disabled = $options.disabled || argDisabled;
 
-			const attach = createAttach();
 			attach('click', () => {
 				if (disabled) return;
 				value.update((v) => {