+++ src/lib/internal/actions/click-outside/clickOutside.types.ts
@@ -1,3 +1,7 @@
+// Modified from Grail UI v0.9.6 (2023-06-10)
+// Source: https://github.com/grail-ui/grail-ui
+// https://github.com/grail-ui/grail-ui/tree/master/packages/grail-ui/src/clickOutside/clickOutside.types.ts
+
 import type { Readable } from 'svelte/store';
 
 export type ClickOutsideConfig = {