+++ src/routes/docs/builders/select/preview.svelte
@@ -2,6 +2,7 @@
 	import { createSelect } from '$lib';
 	import { Docs } from '$routes/(components)';
 	import Check from '~icons/lucide/check';
+	import ChevronDown from '~icons/lucide/chevron-down';
 
 	const { selectedText, trigger, menu, option, isSelected } = createSelect();
 
@@ -12,27 +13,17 @@
 </script>
 
 <Docs.PreviewWrapper>
-	<button
-		class="rounded-md bg-white px-4 py-1 text-magnum-700 outline-none
-	hover:opacity-75 focus:ring focus:ring-magnum-400"
-		{...$trigger()}
-	>
+	<button class="trigger" {...$trigger} aria-label="Food">
 		{$selectedText || 'Select an option'}
+
+		<ChevronDown />
 	</button>
 
-	<ul
-		class="z-10 flex max-h-[300px] min-w-[200px] flex-col gap-2 overflow-y-auto rounded-md
-	bg-white p-1 lg:max-h-none"
-		{...$menu}
-	>
+	<ul class="menu" {...$menu}>
 		{#each Object.entries(options) as [key, arr]}
-			<li class="py-1 pl-4 pr-4 font-bold capitalize text-neutral-800">{key}</li>
+			<li class="label">{key}</li>
 			{#each arr as item}
-				<li
-					class="relative cursor-pointer rounded-md py-1 pl-8 pr-4 text-neutral-800 outline-none
-					hocus:bg-magnum-100 hocus:text-magnum-700"
-					{...$option({ value: item })}
-				>
+				<li class="option" {...$option({ value: item })}>
 					{#if $isSelected(item)}
 						<div class="check">
 							<Check />
@@ -46,8 +37,23 @@
 </Docs.PreviewWrapper>
 
 <style lang="postcss">
+	.label {
+		@apply py-1 pl-4 pr-4 font-semibold capitalize text-neutral-800;
+	}
+	.menu {
+		@apply z-10 flex max-h-[300px] flex-col gap-2 overflow-y-auto;
+		@apply rounded-md bg-white p-1 lg:max-h-none;
+	}
+	.option {
+		@apply relative cursor-pointer rounded-md py-1 pl-8 pr-4 text-neutral-800;
+		@apply outline-none hocus:bg-magnum-100 hocus:text-magnum-700;
+	}
+	.trigger {
+		@apply flex h-10 w-[180px] items-center justify-between rounded-md bg-white px-3;
+		@apply py-2 text-magnum-700 outline-none hover:opacity-75 focus:ring focus:ring-magnum-400;
+	}
 	.check {
-		@apply absolute left-2 top-1/2  text-magnum-500;
+		@apply absolute left-2 top-1/2 text-magnum-500;
 		translate: 0 calc(-50% + 1px);
 	}
 </style>